Create Express Myself
My work looks like it is a square that keeps superimposing and rotating, and small circles are shot from it, then piled up on the background. For this animation code, I have two different explanations, emotions and experience.
First, for emotion, this is a bad meaning. I want to express my pressure during the epidemic. The rotate rectangle could be the pages of the calendar. And infinites small circles that from the rotate rectangle is my pressure. This means that my mood will be affected by different pressures everyday. They could mean unable to go out, unable to shop, boring studying and difficult homework. Also, the monotonous dark color and fast rotation also means my emotion is getting terrible during these days. 
Second is the accumulation of knowledge, which is a good meaning, this is an explanation about experience. Because of COVID-19, I have fallen into the interest of reading during these days. Every page I read is to accumulate knowledge, and those circles are the knowledge that I have accumulated.